+++ Academy/PrivateForm.cs
@@ -46,7 +46,7 @@ namespace Academy
         private void btnOK_Click(object sender, EventArgs e)
         {
             currentPrivate.member = (Member)cbMember.SelectedItem;
-            currentPrivate.Amount = Convert.ToInt32(txtAmount.Text);
+            currentPrivate.Amount = Convert.ToDecimal(txtAmount.Text);
             currentPrivate.BookedLessons = Convert.ToInt32(txtBooked.Text);
             currentPrivate.DoneLessons = Convert.ToInt32(txtDone.Text);
             currentPrivate.Description = txtDescription.Text;
